Only tag the latest image
I decided not to semver tag images because I don't think it's a good
idea to pin this image. Even though not pinning can prevent breakages,
it can also leave you without security updates and bug fixes. I'm making
a judegement call that the risk of breakage is sufficiently low for this
small image that it's best to use `:latest`.